Moved adapter to separate package
